Moved the level code from the game class

+++ src/graphics/tiled_map.h
@@ -20,7 +20,7 @@ struct Tileset {
 
 struct TiledLayer {
     std::string name{""};
-    std::vector<unsigned int> data;
+    std::vector<unsigned short> data;
     sf::VertexArray vertices;
 };
